# app/docs/swagger.yaml
definitions:
...
  domain.FbDataSource:
    properties:
      config:
        description: 数据源对应的配置项：命名空间、请求配置、连接配置、文件路径、是否配置为外部数据源等
        type: string
      id:
        type: integer
      name:
        description: 数据源名称
        type: string
      sourceType:
        description: '数据源类型: 1-db 2-rest 3-graphql 4-自定义'
        type: integer
      switch:
        description: 开关 0-开 1-关
        type: integer
    type: object
...
host: localhost:9123
info:
  contact: {}
  description: FireBoom-server
  title: FireBoom-server Swagger API
  version: "1.0"
paths:
...
  /api/v1/checkDBConn:
    get:
      consumes:
      - application/json
      description: 检查数据源链接
      parameters:
      - description: 数据源类型
        in: formData
        name: sourceType
        required: true
        type: integer
      - description: 数据源链接
        in: formData
        name: url
        required: true
        type: string
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - datasource
  /api/v1/dataSource:
    get:
      consumes:
      - application/json
      description: 查询数据源信息
      parameters:
      - description: 数据源类型
        in: formData
        name: datasourceType
        required: true
        type: integer
      responses:
        "200":
          description: 查询成功
          schema:
            items:
              $ref: '#/definitions/domain.FbDataSource'
            type: array
        "400":
          description: 查询失败
      tags:
      - datasource
    post:
      consumes:
      - application/json
      description: 存储数据源信息
      parameters:
      - description: 数据源名称
        in: formData
        name: name
        required: true
        type: string
      - description: 数据源类型
        in: formData
        name: sourceType
        required: true
        type: integer
      - description: 数据源配置
        in: formData
        name: config
        required: true
        type: string
      - description: 数据源开关 0-开 1-关
        in: formData
        name: switch
        required: true
        type: integer
      responses:
        "200":
          description: 添加成功
        "400":
          description: 添加失败
      tags:
      - datasource
    put:
      consumes:
      - application/json
      description: 修改数据源信息
      parameters:
      - description: 数据源名称
        in: formData
        name: name
        required: true
        type: string
      - description: 数据源类型
        in: formData
        name: sourceType
        required: true
        type: integer
      - description: 数据源配置
        in: formData
        name: config
        required: true
        type: string
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - datasource
  /api/v1/dataSource/:id:
    delete:
      consumes:
      - application/json
      description: 删除数据源信息
      parameters:
      - description: 数据源id
        in: formData
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - datasource
  /api/v1/dataSource/content/:id:
    put:
      consumes:
      - application/json
      description: 修改数据源内容
      parameters:
      - description: 数据源id
        in: formData
        name: id
        required: true
        type: integer
      - description: 数据源内容
        in: formData
        name: content
        required: true
        type: string
      responses:
        "200":
          description: 修改成功
        "400":
          description: 修改失败
      tags:
      - datasource
  /api/v1/dataSource/import:
    post:
      consumes:
      - application/json
      description: 导入外部数据源信息
      parameters:
      - description: 数据源文件
        in: formData
        name: file
        required: true
        type: string
      responses:
        "200":
          description: 导入成功
        "400":
          description: 导入失败
      tags:
      - datasource
...
  /api/v1/file:
    delete:
      consumes:
      - application/json
      description: 删除文件
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 删除成功
        "400":
          description: 删除失败
      tags:
      - file
...
produces:
- application/json
swagger: "2.0"